Add blob name to metrics
