+++ pyrepositoryminer/metrics/blob/__init__.py
@@ -9,6 +9,7 @@ from pyrepositoryminer.metrics.visitor import TreeVisitor
 @dataclass(frozen=True)
 class BlobMetricOutput(TreeMetricOutput):
     blob_id: str
+    blob_name: str
 
 
 class BlobMetric(TreeVisitor, ABC):
